+++ src/main/java/it/contrader/model/Changes.java
@@ -25,6 +25,8 @@ public class Changes {
 	private String name;
 	private String changes;
 	private String removed;
-	private int user;
+	private Long user;
+	
+	
 	
 }
